Move PythonExecInfo.
