+++ src/client/common/process/pythonDaemon.ts
@@ -6,8 +6,8 @@
 import { ChildProcess } from 'child_process';
 import { MessageConnection, RequestType, RequestType0 } from 'vscode-jsonrpc';
 import { InterpreterInformation } from '../../pythonEnvironments/discovery/types';
+import { PythonExecInfo } from '../../pythonEnvironments/exec';
 import { extractInterpreterInfo } from '../../pythonEnvironments/info';
-import { PythonExecInfo } from '../../pythonEnvironments/types';
 import { traceWarning } from '../logger';
 import { BasePythonDaemon } from './baseDaemon';
 import { PythonEnvInfo } from './internal/scripts';
